More details about the limitations of the existing visual aid software for the blind.  This is to facilitate the IT professionals and webpage designers to develop barrier-free web pages.
Currently, there are two main visual aid softwares for the blind, namely Jaws for Windows and NVDA.  Jaws for Windows is the product of Microsoft and is rather expensive.  NVDA is an open-source software developed by a self-help organization of the visually impaired.  No matter which visual aid software is used, they can just show the texts to the visually impaired users.  If the web pages are designed in graphical form without text description, the visually impaired users will be unable to read it.  For example, for normal users, they can like a post in Facebook by clicking on a graphical button, such as a picture of clapping hands.  However, the visually impaired person will be unable to like the post by using the mouse to click the graphical button because the visual aid software cannot locate the button for him/her.  To solve the problem, a text description of "like" must be added onto the graphical button of clapping hands.
To facilitate better understanding, you are encourage to download NVDA from the following page and then open the NVDA to read a selected webpage.  You may find that some of the buttons or contents cannot be read by the software.  The more contents cannot be read, the more barriers for the visually impaired users exist.
https://www.hkbu.org.hk/tc/service/technology/sub-page?page_id=nvda_modules